setup pages when plugin activated

+++ inc/globaltix-function.php
@@ -1,7 +1,20 @@
 <?php
+
+function check_is_page_exist_by_slug($page_slug) {
+
+    $page = get_page_by_path( $page_slug , OBJECT );
+
+    if ( isset($page) ) {
+      return true;
+    } else {
+      return false;
+    }
+}
+
 function setup_pages() {
   $page_guid = site_url() . "/products";
-  $products_page  = array( 'post_title'     => 'Products',
+  $products_page  = array( 
+    'post_title'     => 'Products',
     'post_type'      => 'page',
     'post_name'      => 'products',
     'post_content'   => '',
@@ -10,13 +23,18 @@ function setup_pages() {
     'ping_status'    => 'closed',
     'post_author'    => 1,
     'menu_order'     => 0,
-    'guid'           => $page_guid );
+    'guid'           => $page_guid 
+  );
 
-  $page_id = wp_insert_post( $products_page, FALSE );
-  update_option( 'page_archive_product_id', $page_id, 'yes' );
+  if( check_is_page_exist_by_slug( 'products' ) == false ) {
+    $page_id = wp_insert_post( $products_page, FALSE );
+    update_option( 'page_archive_product_id', $page_id, 'yes' );
+  }
+  
 
   $ticket_page_guid = site_url() . "/tickettype";
-  $ticket_page  = array( 'post_title'     => 'TicketType',
+  $ticket_page  = array( 
+    'post_title'     => 'TicketType',
     'post_type'      => 'page',
     'post_name'      => 'tickettype',
     'post_content'   => '',
@@ -25,10 +43,13 @@ function setup_pages() {
     'ping_status'    => 'closed',
     'post_author'    => 1,
     'menu_order'     => 0,
-    'guid'           => $ticket_page_guid );
-
-  $ticket_id = wp_insert_post( $ticket_page, FALSE );
-  update_option( 'page_ticket_type_id', $page_id, 'yes' );
+    'guid'           => $ticket_page_guid 
+  );
+  
+  if( check_is_page_exist_by_slug( 'tickettype' ) == false ) {
+    $ticket_id = wp_insert_post( $ticket_page, FALSE );
+    update_option( 'page_ticket_type_id', $ticket_id, 'yes' );
+  }
 
 }
